+++ picasso/src/main/java/com/squareup/picasso/ContactsPhotoBitmapHunter.java
@@ -1,15 +1,35 @@
+/*
+ * Copyright (C) 2013 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.squareup.picasso;
 
+import android.annotation.TargetApi;
+import android.content.ContentResolver;
 import android.content.Context;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
+import android.net.Uri;
+import android.os.Build;
+import android.provider.ContactsContract;
 import java.io.IOException;
 import java.io.InputStream;
 
+import static android.provider.ContactsContract.Contacts.openContactPhotoInputStream;
 import static com.squareup.picasso.Request.LoadedFrom.DISK;
-import static com.squareup.picasso.Utils.calculateInSampleSize;
 
-class ContactsPhotoBitmapHunter extends StreamBitmapHunter {
+class ContactsPhotoBitmapHunter extends BitmapHunter {
 
   final Context context;
 
@@ -19,14 +39,41 @@ class ContactsPhotoBitmapHunter extends StreamBitmapHunter {
     this.context = context;
   }
 
-  @Override InputStream getInputStream() throws IOException {
-    return Utils.getContactPhotoStream(context.getContentResolver(), uri);
+  @Override Bitmap decode(Uri uri, PicassoBitmapOptions options) throws IOException {
+    InputStream is = null;
+    try {
+      is = getInputStream();
+      return decodeStream(is, options);
+    } finally {
+      if (is != null) {
+        Utils.closeQuietly(is);
+      }
+    }
   }
 
-  @Override Bitmap decodeStream(InputStream stream, PicassoBitmapOptions options)
-      throws IOException {
+  @Override Request.LoadedFrom getLoadedFrom() {
+    return DISK;
+  }
+
+  private InputStream getInputStream() throws IOException {
+    ContentResolver contentResolver = context.getContentResolver();
+    Uri uri = this.uri;
+    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.ICE_CREAM_SANDWICH) {
+      if (uri.toString().startsWith(ContactsContract.Contacts.CONTENT_LOOKUP_URI.toString())) {
+        uri = ContactsContract.Contacts.lookupContact(contentResolver, uri);
+        if (uri == null) {
+          return null;
+        }
+      }
+      return openContactPhotoInputStream(contentResolver, uri);
+    } else {
+      return ContactPhotoStreamIcs.get(contentResolver, uri);
+    }
+  }
+
+  private Bitmap decodeStream(InputStream stream, PicassoBitmapOptions options) throws IOException {
     if (options != null && options.inJustDecodeBounds) {
-      InputStream is = null;
+      InputStream is = getInputStream();
       try {
         BitmapFactory.decodeStream(is, null, options);
       } finally {
@@ -37,7 +84,10 @@ class ContactsPhotoBitmapHunter extends StreamBitmapHunter {
     return BitmapFactory.decodeStream(stream, null, options);
   }
 
-  @Override Request.LoadedFrom getLoadedFrom() {
-    return DISK;
+  @TargetApi(Build.VERSION_CODES.ICE_CREAM_SANDWICH)
+  private static class ContactPhotoStreamIcs {
+    static InputStream get(ContentResolver contentResolver, Uri uri) {
+      return openContactPhotoInputStream(contentResolver, uri, true);
+    }
   }
 }
