+++ picasso/src/main/java/com/squareup/picasso/ContentProviderBitmapHunter.java
@@ -1,11 +1,32 @@
+/*
+ * Copyright (C) 2013 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.squareup.picasso;
 
+import android.content.ContentResolver;
 import android.content.Context;
+import android.database.Cursor;
 import android.graphics.Bitmap;
 import android.net.Uri;
+import android.provider.MediaStore;
 import java.io.IOException;
 
 class ContentProviderBitmapHunter extends ContentStreamBitmapHunter {
+
+  private static final String[] CONTENT_ORIENTATION =
+      new String[] {MediaStore.Images.ImageColumns.ORIENTATION};
 
   ContentProviderBitmapHunter(Context context, Picasso picasso, Dispatcher dispatcher,
       Request request) {
@@ -13,7 +34,27 @@ class ContentProviderBitmapHunter extends ContentStreamBitmapHunter {
   }
 
   @Override Bitmap decode(Uri uri, PicassoBitmapOptions options) throws IOException {
-    options.exifRotation = Utils.getContentProviderExifRotation(context.getContentResolver(), uri);
+    if (options != null) {
+      options.exifRotation = getContentProviderExifRotation(context.getContentResolver(), uri);
+    }
     return super.decode(uri, options);
+  }
+
+  static int getContentProviderExifRotation(ContentResolver contentResolver, Uri uri) {
+    Cursor cursor = null;
+    try {
+      cursor = contentResolver.query(uri, CONTENT_ORIENTATION, null, null, null);
+      if (cursor == null || !cursor.moveToFirst()) {
+        return 0;
+      }
+      return cursor.getInt(0);
+    } catch (RuntimeException ignored) {
+      // If the orientation column doesn't exist, assume no rotation.
+      return 0;
+    } finally {
+      if (cursor != null) {
+        cursor.close();
+      }
+    }
   }
 }
