+++ picasso/src/main/java/com/squareup/picasso/ContentStreamBitmapHunter.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright (C) 2013 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.squareup.picasso;
 
 import android.content.ContentResolver;
@@ -9,7 +24,6 @@ import java.io.IOException;
 import java.io.InputStream;
 
 import static com.squareup.picasso.Request.LoadedFrom.DISK;
-import static com.squareup.picasso.Utils.calculateInSampleSize;
 
 class ContentStreamBitmapHunter extends BitmapHunter {
 
@@ -29,7 +43,7 @@ class ContentStreamBitmapHunter extends BitmapHunter {
     return DISK;
   }
 
-  Bitmap decodeContentStream(Uri path, PicassoBitmapOptions options) throws IOException {
+  private Bitmap decodeContentStream(Uri path, PicassoBitmapOptions options) throws IOException {
     ContentResolver contentResolver = context.getContentResolver();
     if (options != null && options.inJustDecodeBounds) {
       InputStream is = null;
