+++ picasso/src/main/java/com/squareup/picasso/FileBitmapHunter.java
@@ -1,9 +1,31 @@
+/*
+ * Copyright (C) 2013 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.squareup.picasso;
 
 import android.content.Context;
 import android.graphics.Bitmap;
+import android.media.ExifInterface;
 import android.net.Uri;
 import java.io.IOException;
+
+import static android.media.ExifInterface.ORIENTATION_NORMAL;
+import static android.media.ExifInterface.ORIENTATION_ROTATE_180;
+import static android.media.ExifInterface.ORIENTATION_ROTATE_270;
+import static android.media.ExifInterface.ORIENTATION_ROTATE_90;
+import static android.media.ExifInterface.TAG_ORIENTATION;
 
 class FileBitmapHunter extends ContentStreamBitmapHunter {
 
@@ -12,7 +34,24 @@ class FileBitmapHunter extends ContentStreamBitmapHunter {
   }
 
   @Override Bitmap decode(Uri uri, PicassoBitmapOptions options) throws IOException {
-    options.exifRotation = Utils.getFileExifRotation(uri.getPath());
+    if (options != null) {
+      options.exifRotation = getFileExifRotation(uri.getPath());
+    }
     return super.decode(uri, options);
+  }
+
+  static int getFileExifRotation(String path) throws IOException {
+    ExifInterface exifInterface = new ExifInterface(path);
+    int orientation = exifInterface.getAttributeInt(TAG_ORIENTATION, ORIENTATION_NORMAL);
+    switch (orientation) {
+      case ORIENTATION_ROTATE_90:
+        return 90;
+      case ORIENTATION_ROTATE_180:
+        return 180;
+      case ORIENTATION_ROTATE_270:
+        return 270;
+      default:
+        return 0;
+    }
   }
 }
