+++ picasso/src/main/java/com/squareup/picasso/ResourceBitmapHunter.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright (C) 2013 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.squareup.picasso;
 
 import android.content.Context;
@@ -8,10 +23,8 @@ import android.net.Uri;
 import java.io.IOException;
 
 import static com.squareup.picasso.Request.LoadedFrom.DISK;
-import static com.squareup.picasso.Utils.calculateInSampleSize;
 
 class ResourceBitmapHunter extends BitmapHunter {
-
   private final int resourceId;
   private final Context context;
 
@@ -29,7 +42,8 @@ class ResourceBitmapHunter extends BitmapHunter {
     return DISK;
   }
 
-  Bitmap decodeResource(Resources resources, int resourceId, PicassoBitmapOptions bitmapOptions) {
+  private Bitmap decodeResource(Resources resources, int resourceId,
+      PicassoBitmapOptions bitmapOptions) {
     if (bitmapOptions != null && bitmapOptions.inJustDecodeBounds) {
       BitmapFactory.decodeResource(resources, resourceId, bitmapOptions);
       calculateInSampleSize(bitmapOptions);
