+++ picasso/src/test/java/com/squareup/picasso/RequestBuilderTest.java
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.squareup.picasso;
 
 import android.R;
